+++ huisbot/Modules/Huis/Rankings.cs
@@ -2,7 +2,6 @@
 using Discord.Interactions;
 using huisbot.Helpers;
 using huisbot.Models.Huis;
-using huisbot.Utilities;
 using Microsoft.Extensions.Configuration;
 
 namespace huisbot.Modules.Huis;
