+++ huisbot/Modules/Huis/TopPlays.cs
@@ -4,7 +4,6 @@ using huisbot.Helpers;
 using huisbot.Models.Huis;
 using huisbot.Models.Osu;
 using huisbot.Models.Persistence;
-using huisbot.Utilities;
 using Microsoft.Extensions.Configuration;
 
 namespace huisbot.Modules.Huis;
