+++ huisbot/Modules/Link.cs
@@ -2,7 +2,6 @@
 using Discord.Interactions;
 using huisbot.Helpers;
 using huisbot.Models.Osu;
-using huisbot.Utilities;
 using Microsoft.Extensions.Configuration;
 
 namespace huisbot.Modules;
