+++ huisbot/Modules/Miscellaneous/DiffAttributes.cs
@@ -4,7 +4,6 @@ using huisbot.Helpers;
 using huisbot.Models.Huis;
 using huisbot.Models.Osu;
 using huisbot.Modules.Huis;
-using huisbot.Utilities;
 
 namespace huisbot.Modules.Miscellaneous;
